Port core project to use modules.
