+++ src/Core/core.tmpl
@@ -8,5 +8,4 @@
 #  include <Windows.h>
 #endif
 
-#include <litefx/containers.hpp>
-#include <litefx/traits.hpp>
+#include <litefx/common.h>
