+++ src/Core/include/litefx/.gitignore
@@ -1,3 +1,3 @@
 *.h
-!containers.h
+!common.h
 !litefx.h
