+++ src/Core/modules/alias.ixx
@@ -1,0 +1,168 @@
+module;
+
+#include <unordered_map>
+#include <queue>
+#include <vector>
+#include <span>
+#include <optional>
+#include <memory>
+#include <variant>
+#include <tuple>
+
+#include "generator.hpp"
+
+export module LiteFX.Core:Alias;
+
+export namespace LiteFX {
+
+	/// <summary>
+	/// Represents a handle type.
+	/// </summary>
+	using Handle = void*;
+
+	/// <summary>
+	/// Represents a dictionary that maps a key to a certain value.
+	/// </summary>
+	/// <typeparam name="TKey">The type of the key.</typeparam>
+	/// <typeparam name="TVal">The type of the value.</typeparam>
+	template<class TKey, class TVal>
+	using Dictionary = std::unordered_map<TKey, TVal>;
+
+	/// <summary>
+	/// Represents a dynamic array.
+	/// </summary>
+	/// <typeparam name="T">The type of the array elements.</typeparam>
+	template<class T>
+	using Array = std::vector<T>;
+
+	/// <summary>
+	/// Represents a queue.
+	/// </summary>
+	/// <typeparam name="T">The type of the queue elements.</typeparam>
+	template<class T>
+	using Queue = std::queue<T>;
+
+	/// <summary>
+	/// Represents a view of an array.
+	/// </summary>
+	/// <typeparam name="T">The type of the array elements.</typeparam>
+	template<class T>
+	using Span = std::span<T>;
+
+	/// <summary>
+	/// Represents an optional value.
+	/// </summary>
+	/// <typeparam name="T">The type of the optional value.</typeparam>
+	template<class T>
+	using Optional = std::optional<T>;
+
+	/// <summary>
+	/// Represents a unique pointer, that expresses exclusive ownership.
+	/// </summary>
+	/// <typeparam name="T">The type of the object the pointer points to.</typeparam>
+	/// <typeparam name="TDeleter">The deleter for the pointed object.</typeparam>
+	template<class T, class TDeleter = std::default_delete<T>>
+	using UniquePtr = std::unique_ptr<T, TDeleter>;
+
+	/// <summary>
+	/// Represents a shared pointer, that expresses non-exclusive ownership.
+	/// </summary>
+	/// <typeparam name="T">The type of the object the pointer points to.</typeparam>
+	template <class T>
+	using SharedPtr = std::shared_ptr<T>;
+
+	/// <summary>
+	/// Represents a weak pointer, that expresses a reference to a shared pointer instance.
+	/// </summary>
+	/// <typeparam name="T">The type of the object the pointer points to.</typeparam>
+	template <class T>
+	using WeakPtr = std::weak_ptr<T>;
+
+	/// <summary>
+	/// Represents a tuple of multiple objects.
+	/// </summary>
+	/// <typeparam name="...T">The types of the objects, contained by the tuple.</typeparam>
+	template <class... T>
+	using Tuple = std::tuple<T...>;
+
+	/// <summary>
+	/// Represents a variant of objects.
+	/// </summary>
+	/// <typeparam name="...T">The types of the objects, that can be contained by the tuple.</typeparam>
+	template <class... T>
+	using Variant = std::variant<T...>;
+
+	/// <summary>
+	/// A switch that can be used to select a callable from a parameter type.
+	/// </summary>
+	template<class... TArgs>
+	struct type_switch : TArgs... {
+		using TArgs::operator()...;
+	};
+
+	/// <summary>
+	/// Represents a copyable and assignable reference wrapper.
+	/// </summary>
+	/// <typeparam name="T">The base type of the reference.</typeparam>
+	template <class T>
+	using Ref = std::reference_wrapper<T>;
+
+	/// <summary>
+	/// Represents a synchronous infinite generator range.
+	/// </summary>
+	/// <typeparam name="T">The type of the generated objects.</typeparam>
+	template <typename T>
+	using Generator = std::generator<T>;
+
+	/// <summary>
+	/// Creates a new unique pointer.
+	/// </summary>
+	/// <typeparam name="T">The type of the object, the pointer points to.</typeparam>
+	/// <returns>A new unique pointer.</returns>
+	template <class T>
+	UniquePtr<T> makeUnique() {
+		return std::make_unique<T>();
+	}
+
+	/// <summary>
+	/// Creates a new unique pointer.
+	/// </summary>
+	/// <typeparam name="T">The type of the object, the pointer points to.</typeparam>
+	/// <returns>A new unique pointer.</returns>
+	template <class T, class... TArgs>
+	UniquePtr<T> makeUnique(TArgs&&... _args) {
+		return std::make_unique<T>(std::forward<TArgs>(_args)...);
+	}
+
+	/// <summary>
+	/// Creates a new shared pointer.
+	/// </summary>
+	/// <typeparam name="T">The type of the object, the pointer points to.</typeparam>
+	/// <returns>A new shared pointer.</returns>
+	template <class T>
+	SharedPtr<T> makeShared() {
+		return std::make_shared<T>();
+	}
+
+	/// <summary>
+	/// Creates a new shared pointer.
+	/// </summary>
+	/// <typeparam name="T">The type of the object, the pointer points to.</typeparam>
+	/// <returns>A new shared pointer.</returns>
+	template <class T, class... TArgs>
+	SharedPtr<T> makeShared(TArgs&&... _args) {
+		return std::make_shared<T>(std::forward<TArgs>(_args)...);
+	}
+
+	/// <summary>
+	/// Transfers a unique pointer to a shared pointer. The unique pointer will be released during this process.
+	/// </summary>
+	/// <typeparam name="T">The type of the object, the pointer points to.</typeparam>
+	/// <param name="ptr">The unique pointer that should be turned into a shared pointer.</param>
+	/// <returns>A new shared pointer.</returns>
+	template <class T>
+	SharedPtr<T> asShared(UniquePtr<T>&& ptr) {
+		return SharedPtr<T>(ptr.release());
+	}
+
+}
