+++ src/Core/modules/core.ixx
@@ -1,0 +1,14 @@
+module;
+
+#include <litefx/core.h>
+
+export module LiteFX.Core;
+
+export import :Alias;
+export import :String;
+export import :Concepts;
+export import :Exceptions;
+export import :Enumerable;
+export import :Pimpl;
+export import :Resource;
+export import :Builder;
