--- src/Core/src/core.cpp
@@ -1,3 +1,0 @@
-#include <litefx/core.h>
-
-using namespace LiteFX;
