+++ src/Tests/Core.Enumerable/unique_ptrs.cpp
@@ -6,7 +6,7 @@ int main(int argc, char* argv[])
 	foos.push_back(makeUnique<Foo>(0));
 	foos.push_back(makeUnique<Foo>(1));
 	foos.push_back(makeUnique<Foo>(2));
-	auto bars = []() -> std::generator<UniquePtr<Bar>> {
+	auto bars = []() -> Generator<UniquePtr<Bar>> {
 		for (int i(3); ; ++i)
 			co_yield makeUnique<Bar>(i);
 	}() | std::views::take(3) | std::views::as_rvalue | std::ranges::to<Array<UniquePtr<Base>>>();
@@ -15,10 +15,10 @@ int main(int argc, char* argv[])
 	Enumerable<UniquePtr<Base>> fooBases = foos | std::views::as_rvalue | std::views::drop(1);
 	Enumerable<UniquePtr<Base>> barBases = bars | std::views::reverse | std::views::as_rvalue | std::views::drop(1);
 
-	if (foos[0] == nullptr || foos[1] != nullptr || foos[2] != nullptr)
+	if (!foos[0] || foos[1] || foos[2])
 		return -1;
 
-	if (bars[0] != nullptr || bars[1] != nullptr || bars[2] == nullptr)
+	if (bars[0] || bars[1] || !bars[2])
 		return -2;
 
 	int i = 1;
